Create  pick-up-and-throw-objects-using-the-mouse
This script allows the player to pick up objects by clicking on them with the mouse, and throw them by releasing the mouse button. When an object is picked up, it is parented to a special "Holder" part in the character's model and its collision is disabled. When the object is thrown, it is unparented from the holder and its collision is enabled again. The object is also given a force in the direction that the mouse was pointing when it was released, causing it to be thrown in that direction.